Move AssetCommandView from assets to mission
AssetCommandView reads and writes AssetCommand records — a mission
concern. Move it to mission/views.py and its URL to mission/urls.py
(keeping name='assets_command'). Remove the now-unused timezone import
from assets/views.py.
