+++ assets/urls.py
@@ -12,6 +12,5 @@ urlpatterns = [
     re_path(r'^assets/$', views.AssetsView.as_view(), name='assets_view'),
     re_path(r'^assets/(?P<asset_id>\d+)/$', views.AssetView.as_view(), name='asset_view'),
     re_path(r'^assets/(?P<asset_id>\d+)/status/$', views.AssetStatusView.as_view(), name='assets_status'),
-    re_path(r'^assets/(?P<asset_id>\d+)/command/$', views.AssetCommandView.as_view(), name='assets_command'),
     re_path(r'^assets/status/values/$', views.assets_status_value_list, name='asset_status_values_list'),
 ]
