+++ assets/views.py
@@ -4,7 +4,6 @@ Views for assets
 from django.http import JsonResponse
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, render
-from django.utils import timezone
 from django.utils.decorators import method_decorator
 from django.views import View
 
@@ -110,45 +109,6 @@ class AssetsView(View):
         if "application/json" in request.META.get('HTTP_ACCEPT', ''):
             return self.as_json(request)
         return render(request, 'assets/list.html', {})
-
-
-@method_decorator(login_required, name="dispatch")
-@method_decorator(asset_is_operator, name="dispatch")
-class AssetCommandView(View):
-    """
-    View the current asset command
-    """
-    def as_json(self, request, asset):
-        """
-        Return the current asset command as json
-        """
-        asset_command = AssetCommand.last_command_for_asset_to_json(asset)
-        data = {
-            'command': asset_command,
-        }
-        return JsonResponse(data)
-
-    def get(self, request, asset):
-        """
-        Get the current asset command
-        """
-        return self.as_json(request, asset)
-
-    def post(self, request, asset):
-        """
-        Allow setting the asset command response
-        """
-        command_id = request.POST.get('command_id')
-        asset_command = get_object_or_404(AssetCommand, pk=command_id, asset=asset)
-        type_str = request.POST.get('type')
-        message = request.POST.get('message')
-        if asset_command.responded_at is None:
-            asset_command.responded_at = timezone.now()
-            asset_command.responded_by = request.user
-            asset_command.response_message = message
-            asset_command.response_type = type_str
-            asset_command.save()
-        return self.as_json(request, asset)
 
 
 @method_decorator(login_required, name='dispatch')
